Add cache control "Expires" header to server
Added header:
  Expires: Tue, 19 Jan 2038 03:14:06 GMT
which tells the web browser it is OK to cache the page.  This makes things
more efficient and smoother updating (e.g. the on.png/off.png graphics).

Note that this was done in makefsdata since we disable (the default)
DYNAMIC_HTTP_HEADERS in the httpd.c server code.

Signed-off-by: Gerald Van Baren <gerald.vanbaren@cri.us.com>