Improve config_form CGI, add HTML headers to CGI pages
Return a simple page from the configuration update form CGI to give the
user feedback on update and to cause the web browser to refresh the page
when the user clicks the "OK" button.

Also added HTML headers to the pages to make it more proper, adds cache
control hints for the browser.

Signed-off-by: Gerald Van Baren <gerald.vanbaren@cri.us.com>